+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-CFJpFLCI.js
@@ -1,8 +1,0 @@
-import{c as r,s as e}from"./flowDiagram-HODETNUW-DpEZseFe.js";import{_ as a}from"./mermaid.core-Qw7VzoVW.js";import"./chunk-5VM5RSS4-DoOVpFzW.js";import"./chunk-XXDRQBXY-BQ4gbHMp.js";import"./chunk-POPQ4Y6H-DTo-9Ml9.js";import"./chunk-F27PBJKO-BMlcE6kW.js";import"./channel-DuA204Nz.js";import"./index-CaAwdVlM.js";var o=a(t=>`${e(t)}
-  .swimlane.cluster rect {
-    stroke: ${t.clusterBorder} !important;
-  }
-  [data-look="neo"].cluster rect {
-    filter: none;
-  }
-`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};